new_requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
attrs==25.3.0
av==15.0.0
backoff==2.2.1
certifi==2025.8.3
charset-normalizer==3.4.2
coloredlogs==15.0.1
contourpy==1.3.3
ctranslate2==4.6.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
faster-whisper==1.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inference-sdk==0.51.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
moviepy==2.2.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.22.1
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.1
pillow==11.3.0
proglog==0.1.12
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
PyAudio==0.2.14
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
supervision==0.25.1
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.0
triton==3.4.0
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.176
ultralytics-thop==2.0.15
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
